requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
pyyaml>=6.0.1
